app/requirements.txt
gradio>=5.0.0
databricks-sdk>=0.18.0
requests>=2.31.0
